requirements.txt
fastapi>=0.105.0
uvicorn[standard]>=0.24.0
httpx>=0.25.2
python-dotenv>=1.0.0
openai>=1.2.4,<2.0.0
pydantic>=2.5.2
chainlit>=1.0.0
plotly>=5.18.0
pandas>=2.1.3
jinja2>=3.1.2
aiofiles>=23.2.1